+++ packages/@codotype/ui/src/components/BlueprintEditor/components/schema/SchemaDestroyButton.vue
@@ -1,25 +1,66 @@
 <template>
-  <a
-    id="schema-destroy-button"
-    @click="$store.commit('editor/schema/modals/destroy/showing', true)"
-  >
-    <i class="far fa-fw fa-trash-alt"></i>
-  </a>
+  <span>
+    <!-- <HelpPopover
+      v-if="!hasReverseRelations"
+      target="schema-destroy-button"
+      placement="left"
+      :triggers="[]"
+      content="Remove Schema"
+    /> -->
+    <b-button
+      v-if="hasReverseRelations"
+      class="destroy-button hasReverse"
+      variant="link"
+      v-b-tooltip.hover.left
+      title='Cannot delete Schema with incoming relations'
+    >
+      <i class="far fa-fw fa-trash-alt" />
+    </b-button>
+    <b-button
+      v-else
+      id="schema-destroy-button"
+      class="destroy-button"
+      variant="link"
+      v-b-tooltip.hover.left
+      title='Remove Schema'
+      @click="$store.commit('editor/schema/modals/destroy/showing', true)"
+    >
+      <i class="far fa-fw fa-trash-alt" />
+    </b-button>
+  </span>
 </template>
 
 <script>
+import { mapGetters } from 'vuex'
+import HelpPopover from '../../../HelpPopover'
+
 export default {
-  name: 'SchemaDestroyButton'
+  name: 'SchemaDestroyButton',
+  components: {
+    HelpPopover
+  },
+  computed: {
+    ...mapGetters({
+      inflatedSelectedModel: 'editor/schema/inflatedSelectedModel'
+    }),
+    hasReverseRelations() {
+      return this.inflatedSelectedModel.reverse_relations.length > 0
+    }
+  }
 }
 </script>
 
 <style lang="sass" scoped>
   @import '../../../../sass/vendor.sass'
 
-  a#schema-destroy-button
+  .destroy-button
     transition: color 0.25s ease-in-out
     cursor: pointer
     color: $gray-500
+    &.hasReverse
+      &:hover
+        color: $gray-500 !important
+        cursor: not-allowed
     &:hover
       color: $red
 </style>