+++ packages/@codotype/ui/src/components/BlueprintEditor/store/index.js
@@ -2,26 +2,27 @@ import schemaModule from './schemaModule'
 import projectModule from './projectModule'
 import helpModule from './helpModule'
 import modalModule from '../../../store/lib/modalModule'
-import { DEFAULT_USER_SCHEMA } from '@codotype/types/lib/default_user'
 const DownloadFile = require('downloadjs')
 
 export default {
   namespaced: true,
   actions: {
     created ({ getters, commit, dispatch }) {
-      // if (!getters['schema/collection/first']) {
-      //   commit('schema/collection/items', [Object.assign({}, DEFAULT_USER_SCHEMA)])
-      // }
-      // dispatch('schema/selectModel', getters['schema/collection/first'])
+      // TODO - this should be combined with the logic in the ConfigurationEditor store
+      // These two modules should be combined into the (upcoming) ProjectEditor store
     },
     reset ({ getters, commit, dispatch }) {
       dispatch('project/reset')
-      commit('schema/collection/items', [Object.assign({}, DEFAULT_USER_SCHEMA)])
+      // TODO - pull default schemas here from generator
+      // TODO - pull default schemas here from generator
+      // TODO - pull default schemas here from generator
+      commit('schema/collection/items', [])
       dispatch('schema/selectModel', getters['schema/collection/first'])
     },
     import ({ getters, commit, dispatch }, blueprint) {
       commit('project/label', blueprint.label)
       commit('project/identifier', blueprint.identifier)
+      commit('project/class_name', blueprint.class_name)
       commit('schema/collection/items', blueprint.schemas)
       dispatch('schema/selectModel', getters['schema/collection/first'])
     },
@@ -35,7 +36,7 @@ export default {
       return {
         label: state.project.label || 'Project Name',
         identifier: state.project.identifier || 'project_name',
-        schemas: state.schema.collection.items
+        schemas: state.schema.collection.items.filter(s => s.attributes.length > 0 || s.relations.length > 0)
       }
     }
   },
@@ -47,9 +48,8 @@ export default {
       namespaced: true,
       modules: {
         import: modalModule(),
-        export: modalModule()
+        export: modalModule(),
       }
     }
-
   }
 }