+++ packages/@codotype/ui/src/components/BlueprintEditor/store/localStorageMediator.js
@@ -1,19 +1,25 @@
 export default function configureMediator (store) {
 
-  // Attempt to load localStorage.schemas
+  // Attempt to load localStorage.schemas && localStorage.projectLabel
   try {
+    // Loads the projectLabel
+    let { projectLabel } = window.localStorage
+    if (projectLabel) {
+      store.dispatch('editor/project/setLabel', projectLabel)
+    }
+
+    // Loads the schemas
     let loaded = JSON.parse(window.localStorage.schemas)
     store.commit('editor/schema/collection/items', loaded)
-    // console.log('LOADED')
   } catch {
     // console.log('ERROR LOADING FROM LOCAL STORAGE')
   }
 
   store.subscribe(({ type, payload }) => {
+    if (type === 'editor/project/label') {
+      window.localStorage.setItem('projectLabel', payload)
+    }
     if (type === 'editor/schema/collection/items') {
-      // console.log('UPDATE STORAGE HERE')
-      // console.log(type)
-      // console.log(payload)
       const stringified = JSON.stringify(payload)
       window.localStorage.setItem('schemas', stringified)
     }