+++ packages/@codotype/ui/src/components/BlueprintEditor/store/projectModule.js
@@ -1,14 +1,31 @@
 import sanitizeLabel from '@codotype/util/lib/sanitizeLabel'
 import inflateMeta from '@codotype/util/lib/inflateMeta'
+import modalModule from '../../../store/lib/modalModule'
 
 export default {
   namespaced: true,
   state: {
-    label: '',
-    identifier: '',
-    class_name: ''
+    editLabel: '',
+    editIdentifier: '',
+    label: 'Project Name',
+    identifier: 'project_name',
+    class_name: 'ProjectName'
   },
   actions: {
+    edit ({ state, commit, dispatch }) {
+      dispatch('setEditLabel', state.label)
+      commit('modals/form/showing', true)
+    },
+    submitForm ({ state, commit, dispatch }) {
+      dispatch('setLabel', state.editLabel)
+      commit('modals/form/showing', false)
+    },
+    setEditLabel ({ commit }, label) {
+      const sanitizedLabel = sanitizeLabel(label)
+      const { identifier } = inflateMeta(sanitizedLabel)
+      commit('editLabel', sanitizedLabel)
+      commit('editIdentifier', identifier)
+    },
     setLabel ({ commit }, label) {
       const sanitizedLabel = sanitizeLabel(label)
       const { identifier, class_name } = inflateMeta(sanitizedLabel)
@@ -23,17 +40,29 @@ export default {
     }
   },
   mutations: {
+    editLabel (state, label) {
+      state.editLabel = label
+    },
+    editIdentifier (state, identifier) {
+      state.editIdentifier = identifier
+    },
     label (state, label) {
       state.label = label
     },
     identifier (state, identifier) {
       state.identifier = identifier
     },
-    identifier (state, class_name) {
+    class_name (state, class_name) {
       state.class_name = class_name
     }
   },
   getters: {
+    editLabel: state => {
+      return state.editLabel
+    },
+    editIdentifier: state => {
+      return state.editIdentifier
+    },
     label: state => {
       return state.label
     },
@@ -43,8 +72,16 @@ export default {
     class_name: state => {
       return state.class_name
     },
-    enableSubmit: state => {
-      return state.identifier !== "" && state.identifier !== "" && state.label.length > 2
+    disableSubmit: state => {
+      return state.editIdentifier === "" || state.editIdentifier === "" || state.editLabel.length < 2
+    }
+  },
+  modules: {
+    modals: {
+      namespaced: true,
+      modules: {
+        form: modalModule(),
+      }
     }
   }
 }